backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.3
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
python-multipart==0.0.22
httpx==0.26.0
aiofiles==23.2.1
